remove access control -> self governance and mono contract design for demo
