+++ kairos-contracts/demo-contract-tests/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "deposit-contract-tests"
+name = "demo-contract-tests"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
