+++ kairos-contracts/demo-contract/contract/src/constants.rs
@@ -1,0 +1,17 @@
+pub const KAIROS_DEPOSIT_CONTRACT_NAME: &str = "kairos_demo_contract";
+
+//#[allow(dead_code)]
+pub const KAIROS_DEPOSIT_CONTRACT_PACKAGE: &str = "kairos_contract_package";
+pub const KAIROS_DEPOSIT_CONTRACT_UREF: &str = "demo_contract";
+
+pub const KAIROS_LAST_PROCESSED_DEPOSIT_COUNTER: &str = "last_processed_deposit_counter";
+pub const KAIROS_DEPOSIT_PURSE: &str = "kairos_deposit_purse";
+
+pub const RUNTIME_ARG_TEMP_PURSE: &str = "temp_purse";
+
+//#[allow(dead_code)]
+pub const RUNTIME_ARG_AMOUNT: &str = "amount";
+
+pub const EP_INIT_NAME: &str = "init";
+pub const EP_GET_PURSE_NAME: &str = "get_purse";
+pub const EP_DEPOSIT_NAME: &str = "deposit";
