+++ kairos-contracts/demo-contract/contract/src/utils/events.rs
