+++ kairos-contracts/demo-contract/deposit-session/src/main.rs
@@ -13,7 +13,7 @@ use casper_types::{runtime_args, ContractHash, RuntimeArgs, URef, U512};
 
 #[no_mangle]
 pub extern "C" fn call() {
-    let contract_hash: ContractHash = runtime::get_named_arg("deposit_contract");
+    let contract_hash: ContractHash = runtime::get_named_arg("demo_contract");
     let amount: U512 = runtime::get_named_arg("amount");
     let source: URef = account::get_main_purse();
     // create a temporary purse that can be passed to the deposit contract
