+++ kairos-contracts/demo-contract/malicious-reader/Cargo.toml
