+++ kairos-contracts/demo-contract/malicious-reader/src/main.rs
