+++ kairos-contracts/demo-contract/malicious-session/Cargo.toml
