+++ kairos-contracts/demo-contract/malicious-session/src/main.rs
@@ -11,7 +11,7 @@ use casper_types::{runtime_args, ContractHash, RuntimeArgs, URef, U512};
 
 #[no_mangle]
 pub extern "C" fn call() {
-    let contract_hash: ContractHash = runtime::get_named_arg("deposit_contract");
+    let contract_hash: ContractHash = runtime::get_named_arg("demo_contract");
     let amount: U512 = runtime::get_named_arg("amount");
     let destination_purse: URef = account::get_main_purse();
     let borrowed_contract_purse: URef =
