--- kairos-contracts/deposit-contracts/contract/src/constants.rs
@@ -1,23 +1,0 @@
-pub const KAIROS_DEPOSIT_CONTRACT_NAME: &str = "kairos_deposit_contract";
-// TODO why is this unused?
-#[allow(dead_code)]
-pub const KAIROS_DEPOSIT_CONTRACT_PACKAGE: &str = "deposit_contract_package";
-pub const KAIROS_DEPOSIT_CONTRACT: &str = "deposit_contract";
-
-pub const KAIROS_LAST_PROCESSED_DEPOSIT_COUNTER: &str = "last_processed_deposit_counter";
-pub const KAIROS_DEPOSIT_PURSE: &str = "kairos_deposit_purse";
-
-pub const RUNTIME_ARG_TEMP_PURSE: &str = "temp_purse";
-// TODO why is this unused?
-#[allow(dead_code)]
-pub const RUNTIME_ARG_DEST_PURSE: &str = "destination_purse";
-pub const RUNTIME_ARG_AMOUNT: &str = "amount";
-
-pub const SECURITY_BADGES: &str = "security_badges";
-pub const ADMIN_LIST: &str = "admin_list";
-
-pub const EP_INIT_NAME: &str = "init";
-pub const EP_GET_PURSE_NAME: &str = "get_purse";
-pub const EP_DEPOSIT_NAME: &str = "deposit";
-pub const EP_INCR_LAST_PROCESSED_NAME: &str = "incr_last_processed_deposit_counter";
-pub const EP_UPDATE_SECURITY_BADGES_NAME: &str = "update_security_badges";
